:recycle: Fix 캘린더 파일 생성 #6
